+++ viewith-server/src/main/java/com/puremadeleine/viewith/config/auth/SecurityProperties.java
@@ -1,15 +1,16 @@
-package com.puremadeleine.viewith.config;
+package com.puremadeleine.viewith.config.auth;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
+import java.util.Set;
+
 @Getter
 @Setter
 @Component
-@ConfigurationProperties(prefix = "jwt")
-public class JwtProperties {
-    String accessSecretKey;
-    String refreshSecretKey;
+@ConfigurationProperties(prefix = "security")
+public class SecurityProperties {
+    Set<String> excludeUris;
 }
