+++ viewith-server/src/main/java/com/puremadeleine/viewith/config/client/FeignLoggerConfig.java
@@ -1,4 +1,4 @@
-package com.puremadeleine.viewith.config;
+package com.puremadeleine.viewith.config.client;
 
 import feign.Logger;
 import org.springframework.context.annotation.Bean;
