+++ viewith-server/src/main/java/com/puremadeleine/viewith/config/client/KakaoOAuthProperties.java
@@ -1,4 +1,4 @@
-package com.puremadeleine.viewith.config;
+package com.puremadeleine.viewith.config.client;
 
 import lombok.Getter;
 import lombok.Setter;
