+++ viewith-server/src/main/java/com/puremadeleine/viewith/config/jpa/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.puremadeleine.viewith.config;
+package com.puremadeleine.viewith.config.jpa;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
