+++ viewith-server/src/main/java/com/puremadeleine/viewith/repository/client/KakaoOAuthRepository.java
@@ -1,6 +1,6 @@
 package com.puremadeleine.viewith.repository.client;
 
-import com.puremadeleine.viewith.config.FeignLoggerConfig;
+import com.puremadeleine.viewith.config.client.FeignLoggerConfig;
 import com.puremadeleine.viewith.dto.client.AccessTokenResDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 @FeignClient(name = "kakao-oauth", url = "${kakao.oauth.uri}", configuration = FeignLoggerConfig.class)
 public interface KakaoOAuthRepository {
-    @PostMapping(value="/oauth/token", consumes = "application/x-www-form-urlencoded")
+    @PostMapping(value = "/oauth/token", consumes = "application/x-www-form-urlencoded")
     AccessTokenResDto getAccessToken(@RequestParam("grant_type") String grantType,
-                                     @RequestParam("client_id")String clientId,
-                                     @RequestParam("request_uri")String redirectUri,
-                                     @RequestParam("code")String code,
-                                     @RequestParam("client_secret")String clientSecret);
+                                     @RequestParam("client_id") String clientId,
+                                     @RequestParam("request_uri") String redirectUri,
+                                     @RequestParam("code") String code,
+                                     @RequestParam("client_secret") String clientSecret);
 }
